Fix X-Request-Id to be a native string on py2 (#247)
`future.utils.native` doesn't convert unicode to str on Python 2; the
function we need for that is `future.utils.text_to_native_str`.  This
caused requests with non-ASCII bodies to fail due to trying to
concatenate a unicode object with a non-ASCII str.

This is tough to test end-to-end because responses hooks in high enough
to mask the crash in httplib, but testing that all header names are
native strings should at least be enough to stop it regressing.

This bug was introduced in 0.9.6.