Properly serialize the prometheus worker cleanup (#296)
* Properly serialize the prometheus worker cleanup

The child_exit hook now just records the pid that needs cleaning, the
actual work is done in gunicorn's main loop, serialized and safe. To do
this we have to add a fake 'signal' to hook into gunicorns loop, but it
works. Also adds a lock around the metric aggregation writes stage