Refactor the way Talisker keeps sentry client references up to date. (#344)
* Refactor the way Talisker keeps sentry client references up to date.

Move to a property based lookup, rather than tracking/updating actual
references.  This enables a future branch that changes how testing
sentry works.

This is needed because flask/django create/configure new sentry client
instances, using their relative app configuration, when they are
initialised. Talisker has by this point already set up various handlers
(e.g. logging, WSGI), and these have a reference to the client they were
created with (which may be missing application configuration). So we
need to update those to point to the new client instance. This is done
by subclassing and using a property to make the .client attribute
actually be a call to get_client() under the hood.