Refactor internal context API
This refactors the previously multiple bits of request context into
a single object and API, that doesn't not rely on werkzeug.

This cleans things up a lot, provides a much nicer context API to use,
and also allows us to get rid the the request_id module altogether.

It paves the way for a) dropping werkzeug as a required dependency and
b) using contextvars to store this new ContextData object