Better gunicorn timeout handling.
By raising our own exception in Gunicorn's worker_abort handler, we can
catch and handle it just like any other error. This means we get an
error response, logging, and sentry report for the timeout, which is
a massive improvement.

Also includes a few of drive-by tweaks - disable parallel pytest runs in
tox, as it generally confuses things, tweak the error message for
assert_log help to reduce noise, and slight tweaks to the title of the
default error pages.