Support X-Debug header that sends sentry report.
As a precaution, the header is only respected from a trusted ip
address. This provides a secure way to to targeted diagnostic debugging.

This change also reintroduced the X-Sentry-Id header, as I figured out
how to include that when possible. It's not possible to do so for soft
timeouts, as we don't know until after the request is sent. Nor for
errors that occur after headers have been sent.