Require contexts to be explicitly created.
Implicitly creating contexts causes a variety of issues, and was a bit
of a hack when first doing the contextvars work.

This change introduces a NullContext, which is the default unless
you've explicitly created one. This NullContext provides the same API as
a context, but does not store any information.

This fixes a few different issues, from context bleeding and leaking,
through to test fixture issues that clear the context dict.

This is an internal API change, and shouldn't affect users code.