Send view-name (if available) to sentry with soft timeout messages
This should make the default grouping of those events in Sentry more
accurate in absence of other attributes like stacktrace in this event.
Particularly for views with url path args.

This changes both the "message" itself and adds a transaction tag/field
to the event if it's not set (it might not be available in this context
for example with Django when the appropriate middleware had already
cleared the transaction name).