Stop using TaliskerRequestsTransport
At least for now stop using the TaliskerRequestsTransport.

In practice, for flask apps it was not used in the apps/workers because
FlaskSentry was explicitly setting a None 'transport' in the client
kwargs.

This change makes sure that the same default raven (threaded) transport
is used by logging in the main process. This will mitigate the deadlocks
described in
https://github.com/prometheus/client_python/pull/1076#issuecomment-2513729120
because there will be no prometheus instrumentation for sentry requests.